Continued updating SensorThings engine
