+++ core/fixtures/observed_properties_masterlist.yaml
@@ -5088,7 +5088,7 @@
 #     definition:
 #       Geographical land classification- genetically homogeneous territorial
 #       unit within natural boundaries with a certain structure and a certain character
-#       of interrelationships of considered components
+#       of interrelationships of considered engine
 #     description: ""
 #     name: Land classification
 #     person: null
@@ -8473,7 +8473,7 @@
 #       to those of its two vertical neighbors in the periodic table, calcium and barium.
 #       It occurs naturally mainly in the minerals celestine and strontianite, and is
 #       mostly mined from these. While natural strontium is stable, the synthetic 90Sr
-#       isotope is radioactive and is one of the most dangerous components of nuclear
+#       isotope is radioactive and is one of the most dangerous engine of nuclear
 #       fallout, as strontium is absorbed by the body in a similar manner to calcium.
 #       Natural stable strontium, on the other hand, is not hazardous to health.
 #     description: Wikipedia
@@ -8691,7 +8691,7 @@
 #     definition:
 #       Tantalum is a rare, hard, blue-gray, lustrous transition metal that
 #       is highly corrosion-resistant. It is part of the refractory metals group, which
-#       are widely used as minor components in alloys.
+#       are widely used as minor engine in alloys.
 #     description: Wikipedia
 #     name: Tantalum
 #     person: null
