+++ tests/test_sensorthings_endpoints.py
@@ -143,7 +143,7 @@ def test_sensorthings_get_endpoints(
     }, 201),
     ('Observations', [{
         'Datastream': {'@iot.id': '376be82c-b3a1-4d96-821b-c7954b931f94'},
-        'components': ['phenomenonTime', 'result'],
+        'engine': ['phenomenonTime', 'result'],
         'dataArray': [['2023-10-02T10:00:00Z', 24.5], ['2023-10-03T10:00:00Z', 22.8]]
     }], 201)
 ])
